requirements.txt
requests~=2.32.3
PyYAML~=6.0.2
pendulum~=3.0.0
croniter~=5.0.1